refactor(snapp): get files from folder
DropFileArea component should get files just like upload-dialog
